Refactor location of StringFormatter
