+++ janis_core/operators/__init__.py
@@ -1,2 +1,3 @@
 from .operator import *
 from .selectors import *
+from .stringformatter import StringFormatter
